feat: Update Hugo archetypes to remove HRs and change final subsection (#531)
This commit updates the Hugo archetypes to remove the horizontal lines
delineating each section, and also changes the final sections of each
section. The future website theme uses meaningful whitespace for
breaking sections up instead of horizontal lines, so removing it from
the templates prevents more from being added.

"See also" was changed to "Next steps" for two of the archetypes, with
more specific examples for all three. This reflects discussions around
the fact that "See also" doesn't inspire informed content design.

"Next steps" is more appropriate for specific use cases, especially with
the context of the working document - we have a good idea of what the
reader may need to do next, so the title should reflect that.