feat: Add proposals section, and NGINX App Protect WAF Refactor
This commit adds a new documentation folder to the repository, with a
discrete section for proposals. I will be using this folder in the
future to organise more process documentation.

NGINX App Protect WAF has had a lot of fundamental design problems with
how its documentation is structured due to historic changes in the
direction of the product. Since those changes, the documentation has not
changed to account for it.