feat: Update repo READMEs and affected files (#783)
This commit removes the CHANGELOG file from the repository, as it was
not being used. The README and CONTRIBUTING files have been streamlined,
and now more accurately reflect the user flow when contributing.

Some GitHub templates have been updated to reflect these changes, and
also include relative links instead of axiomatic blobs, which will make
finding and updating links easier in the future.