feat: Add OpenTelemetry documentation (#824)
This commit adds OpenTelemetry documentation for NGINX Ingress
Controller, converting an example hidden in the code repository. It adds
it to the "Logging and monitoring" section, where the prior OpenTracing
document has also been moved.

The latter document has had an explicit warning call-out added to the
top directing users to the new document, and the release page now links
directly to the new configuration document for context.

Some page titles and metadata in the section have also been updated as
part of this work, to ensure the user experience is more consistent.

Co-authored-by: yar <y82@users.noreply.github.com>