feat: Update Hugo and PR template guidance (#894)
This commit updates the Hugo guidance to accurately reflect the usage of
our most common shortcodes, and guides people away from the ones that
have been removed or will be removed soon based on Mainframe changes.

The guidance is not fully inclusive: we have not yet determined where
full documentation for all theme elements should live (Such as a
publicly-facing design system), but it should cover the most commonly
used shortcodes.

The pull request template has also been updated in the same manner,
attempting to focus on critical information for contributors. The
checklist was previously quite long: it has been reduced to essential
elements, and reframed as a direct, personal commitment.

The detail underneath the "Proposed changes" has been shifted to inline,
hidden guidance, following a similar pattern in our Hugo archetypes.
Some information from the old checklist has been moved there, but it
otherwise covers the same ground as the old guidance.

It refers to and reinforces our standards asserted by our Git
conventions documentation, and will not create visual noise within the
PR should the contributor choose not to remove it.