feat: Clarify NGINX Ingress Controller install & upgrade documentation (#816)
This commit makes a number of changes to clarify what the precise steps
are to install, upgrade and uninstall NGINX Ingress Controller

- The Install document clarifies necessary steps
- Uninstall steps have been moved to the bottom of the document
- Upgrade instructions have been moved into their own document

The changes made focused on clarifying the critical path for one single
user story, as lots of unnecessary contextual information was being
provided, which was causing confusion.

With this commit, the general reading and execution flow for these
deployment steps become more coherent. Similar changes may be made for
Manifests documentation, if common steps are identified.

As part of the work, a number of include files have been updated or created 
that are shared with NGINX Plus, to ensure the documentation around 
obtaining an NGINX Ingress Controller image with Docker are consistent.

Subsequently, process documentation for managing includes was split from 
the style guide and made into its own file due to the anticipated length 
as reference information.

---------

Co-authored-by: Travis Martin <33876974+travisamartin@users.noreply.github.com>