feat: Add core limitations and missing Docker configuration (#1356)
This commit updates the Markdown linting configuration files, and adds
an explanation of the linting available in the repository to the process
documentation. The linter itself has been used on multiple files also
affected in this commit.

There are a handful of peripheral changes made in the PR:

- Adding the latest catalogue tags to the Hugo archetype metadata
- Fixing a broken link for NGINX Amplify
- Fixing duplicate tabs caused by includes (Closes #1343)
- Adding reload instructions to NGINX configuration (Closes #1346)
- Configuration updates for the repository's automatic labeller

Additionally, a small heading naming issue was fixed in the changelog,
as well as a formatting style based on our Markdown linting rules.